+++ CHANGES.md
@@ -1,7 +1,7 @@
 # 0.14.0
 
 * Add optional options `:input_id` and `:output_id` to `VariableMapping.merge_instructions_for`. You don't need this, though.
-
+* Deprecate the taskWrap extension API in favor of taskWrap.Step(). You will see warnings explaining the update.
 
 # 0.13.0